--- Models/FictionalUser/FictionalUser.cs
@@ -1,7 +1,0 @@
-namespace bexio.net.Models
-{
-    public class FictionalUser : FictionalUserBase
-    {
-        public string id { get; set; }
-    }
-}
