--- Models/Timesheet/TimesheetStatus.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-
-namespace bexio.net.Models
-{
-	public class TimesheetStatus
-	{
-		public string id { get; set; }
-		public string name { get; set; }
-	}
-}
