--- Models/Tracking/TrackingUpdate.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace bexio.net.Models
-{
-	public class TrackingUpdate
-	{
-		public string date { get; set; }
-		public string duration { get; set; }
-	}
-}
