--- Models/User/User.cs
@@ -1,7 +1,0 @@
-namespace bexio.net.Models
-{
-    public class User : UserBase
-    {
-        public string id { get; set; }
-    }
-}
