+++ bexio.net/Models/FictionalUser/FictionalUser.cs
@@ -1,0 +1,7 @@
+namespace bexio.net.Models
+{
+    public class FictionalUser : FictionalUserBase
+    {
+        public string id { get; set; }
+    }
+}
