+++ bexio.net/Models/Timesheet/Timesheet.cs
@@ -1,0 +1,9 @@
+namespace bexio.net.Models
+{
+	public class Timesheet : TimesheetBase
+	{
+		public int id { get; set; }
+		public string date { get; set; }
+		public string duration { get; set; }
+	}
+}
