+++ bexio.net/Models/Timesheet/TimesheetStatus.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace bexio.net.Models
+{
+	public class TimesheetStatus
+	{
+		public string id { get; set; }
+		public string name { get; set; }
+	}
+}
