+++ bexio.net/Models/Tracking/TrackingUpdate.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using Newtonsoft.Json;
+
+namespace bexio.net.Models
+{
+	public class TrackingUpdate
+	{
+		public string date { get; set; }
+		public string duration { get; set; }
+	}
+}
