+++ bexio.net/Models/User/User.cs
@@ -1,0 +1,7 @@
+namespace bexio.net.Models
+{
+    public class User : UserBase
+    {
+        public string id { get; set; }
+    }
+}
