+++ bexio.net/Models/User/UserBase.cs
@@ -1,0 +1,12 @@
+namespace bexio.net.Models
+{
+	public class UserBase
+	{
+		public string salutation_type { get; set; }
+		public string firstname { get; set; }
+		public string lastname { get; set; }
+		public string email { get; set; }
+		public string is_superadmin { get; set; }
+		public string is_accountant { get; set; }
+	}
+}
